refactor(reports): Reorganized database models
