+++ reports/tests/unittests/test_get_reports_by_day.py
@@ -1,4 +1,3 @@
-from datetime import date
 from unittest import mock
 
 from pymongo.errors import ServerSelectionTimeoutError
