+++ reports/tests/unittests/test_patch_report.py
@@ -4,7 +4,7 @@ from unittest import mock
 from db_handler import DocumentNotFound
 from pymongo.errors import DuplicateKeyError, ServerSelectionTimeoutError
 
-from reports.database import Report
+from reports.database.models import Report
 from reports.schemas import ReportUpdate
 from .. import utils
